Architecture - move simpler classes out to own files Change TransmissionUnit to TransactionUnit
