+++ FileTransferCS/Received/test.txt
@@ -6,7 +6,6 @@
 123456789ABCDEF
 123456789ABCDEF
 123456789ABCDEF
-§§§§***************
 ***************
 ***************
 ***************
@@ -14,4 +13,5 @@
 ***************
 ***************
 ***************
-§§§§<end>§§§§ннн
+***************
+<end>
